# agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autonomous Hydroponic Plant API",
    "description": "REST API to control and monitor the ZombiePlant V1.0 hydroponic system.",
    "version": "1.1"
  },
  "paths": {
...
    "/control/pump": {
      "post": {
        "tags": [
          "Control"
        ],
        "summary": "Control Pump",
        "description": "Manually dispense from a specific pump for a given duration.",
        "operationId": "control_pump_control_pump_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PumpCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PumpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/control/ac_relay": {
      "post": {
        "tags": [
          "Control"
        ],
        "summary": "Control Ac Relay",
        "description": "Turn AC Power 'on' or 'off' for devices like the main grow light.",
        "operationId": "control_ac_relay_control_ac_relay_post",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RelayState",
              "description": "The desired state for the AC power."
            },
            "description": "The desired state for the AC power."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ACRelayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sensors/camera/capture": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Capture a standard photo",
        "description": "Captures an image with default camera settings.",
        "operationId": "capture_photo_sensors_camera_capture_get",
        "parameters": [
          {
            "name": "autofocus_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Autofocus mode: default, manual, continuous",
              "title": "Autofocus Mode"
            },
            "description": "Autofocus mode: default, manual, continuous"
          },
          {
            "name": "lens_position",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Lens position for manual focus (0.0 - infinity)",
              "title": "Lens Position"
            },
            "description": "Lens position for manual focus (0.0 - infinity)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {}
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/camera/plant": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Capture a photo with grow light ON",
        "description": "Captures an image with the main light (AC Relay) turned ON.\nTuned for bright light: Lowers EV and Saturation.",
        "operationId": "capture_plant_photo_sensors_camera_plant_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {}
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/microphone/record": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "summary": "Record a short audio clip",
        "description": "Records an audio clip, useful for detecting pump/system noises.",
        "operationId": "record_audio_sensors_microphone_record_get",
        "parameters": [
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "exclusiveMinimum": 0,
              "description": "Recording duration in seconds.",
              "default": 5,
              "title": "Duration"
            },
            "description": "Recording duration in seconds."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "audio/wav": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ACRelayResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "ac_power": {
            "$ref": "#/components/schemas/RelayState"
          }
        },
        "type": "object",
        "required": [
          "ac_power"
        ],
        "title": "ACRelayResponse"
      },
      "CameraErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "const": "Capture failed",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "CameraErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PumpCommand": {
        "properties": {
          "pump_id": {
            "$ref": "#/components/schemas/PumpID",
            "description": "The unique identifier for the pump."
          },
          "duration": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Duration",
            "description": "Duration in seconds for which to run the pump.",
            "example": 5.5
          }
        },
        "type": "object",
        "required": [
          "pump_id",
          "duration"
        ],
        "title": "PumpCommand"
      },
      "PumpID": {
        "type": "string",
        "enum": [
          "water_out",
          "water_in",
          "flora_micro",
          "flora_gro",
          "flora_bloom"
        ],
        "title": "PumpID"
      },
      "PumpResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "success",
            "title": "Status",
            "default": "success"
          },
          "pump": {
            "$ref": "#/components/schemas/PumpID"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          }
        },
        "type": "object",
        "required": [
          "pump",
          "duration"
        ],
        "title": "PumpResponse"
      },
      "RelayState": {
        "type": "string",
        "enum": [
          "on",
          "off"
        ],
        "title": "RelayState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}